+++ src/gamspy/math/trigonometric.py
@@ -27,7 +27,7 @@ def cos(x: int | float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = cos(np.pi)
     >>> r.toValue()
-    -1.0
+    np.float64(-1.0)
 
     """
     return expression.Expression(None, MathOp("cos", (x,)), None)
@@ -49,7 +49,7 @@ def cosh(x: int | float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = cosh(0)
     >>> r.toValue()
-    1.0
+    np.float64(1.0)
 
     """
     return expression.Expression(None, MathOp("cosh", (x,)), None)
@@ -72,7 +72,7 @@ def sin(x: float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = sin(np.pi/2)
     >>> r.toValue()
-    1.0
+    np.float64(1.0)
 
     """
     return expression.Expression(None, MathOp("sin", (x,)), None)
@@ -94,7 +94,7 @@ def sinh(x: float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = sinh(0)
     >>> r.toValue()
-    0.0
+    np.float64(0.0)
 
     """
     return expression.Expression(None, MathOp("sinh", (x,)), None)
@@ -117,7 +117,7 @@ def tan(x: float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = tan(np.pi/4)
     >>> round(r.toValue(), 2)
-    1.0
+    np.float64(1.0)
 
     """
     return expression.Expression(None, MathOp("tan", (x,)), None)
@@ -139,7 +139,7 @@ def tanh(x: float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = tanh(0)
     >>> r.toValue()
-    0.0
+    np.float64(0.0)
 
     """
     return expression.Expression(None, MathOp("tanh", (x,)), None)
@@ -161,7 +161,7 @@ def acos(x: float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = acos(1)
     >>> r.toValue()
-    0.0
+    np.float64(0.0)
 
     """
     return expression.Expression(None, MathOp("arccos", (x,)), None)
@@ -183,7 +183,7 @@ def asin(x: float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = asin(0)
     >>> r.toValue()
-    0.0
+    np.float64(0.0)
 
     """
     return expression.Expression(None, MathOp("arcsin", (x,)), None)
@@ -205,7 +205,7 @@ def atan(x: float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = atan(0)
     >>> r.toValue()
-    0.0
+    np.float64(0.0)
 
     """
     return expression.Expression(None, MathOp("arctan", (x,)), None)
@@ -227,7 +227,7 @@ def atan2(y: int | float | Symbol, x: int | float | Symbol) -> Expression:
     >>> r = Parameter(m, "r")
     >>> r[...] = atan2(1,1)
     >>> r.toValue()
-    0.7853981633974483
+    np.float64(0.7853981633974483)
 
     """
     return expression.Expression(None, MathOp("arctan2", (y, x)), None)
